requirements.txt
# Core dependencies
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.5.0
pyyaml>=6.0

# RL framework
stable-baselines3>=2.0.0
gymnasium>=0.28.1

# Deep learning
# Using torch>=2.2.0 for torch.from_numpy instead of deprecated torch.tensor
torch>=2.2.0
torch-geometric>=2.3.1

# Quantum simulation
stim>=1.10.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
